+++ python/assassyn/ir/expr/__init__.py
@@ -2,6 +2,7 @@
 
 #pylint: disable=wildcard-import
 from .expr import *
+from .arith import *
 from .intrinsic import Intrinsic, PureIntrinsic, finish, wait_until, assume, barrier
 from .intrinsic import mem_write, send_read_request, send_write_request
 from .intrinsic import has_mem_resp, mem_resp, use_dram
